twister: move platform class out
Move both platform and config parser into own file.
